refactor(tvshow-module): add viewmodel

+++ base/src/main/java/com/godohdev/base/data/model/ResultsItem.kt
@@ -35,7 +35,7 @@ data class ResultsItem(
 	val popularity: Double? = null,
 
 	@field:SerializedName("vote_average")
-	val voteAverage: Int? = null,
+	val voteAverage: Double? = null,
 
 	@field:SerializedName("id")
 	val id: Int? = null,
